# contracts/design.openapi.yaml
openapi: 3.1.0
info:
  title: Imora Design API
  version: 0.2.0
  description: >-
    AI design module — upload, queue, variants, material tags. Owner Abdunazar.

    v0.2.0 adds GET /design/quota, GET /design/requests, the nameUz style field,
    and provider/model/errorMessage/completedAt on DesignRequest.
servers:
  - url: http://localhost:3001/api
tags:
  - name: Design
paths:
  /design/styles:
    get:
      tags: [Design]
      summary: Style library used in prompts
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DesignStyle'
  /design/quota:
    get:
      tags: [Design]
      security: [{ bearerAuth: [] }]
      summary: Remaining free generations, so the UI can warn before upload (AC-11)
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DesignQuota'
  /design/requests:
    get:
      tags: [Design]
      security: [{ bearerAuth: [] }]
      summary: The caller's recent design requests
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DesignRequest'
    post:
      tags: [Design]
      security: [{ bearerAuth: [] }]
      summary: Upload a room photo and enqueue generation (AC-08)
      description: >-
        Returns 202 immediately; poll GET /design/requests/{id} until status is
        done or failed. Accepts JPEG, PNG or WebP up to 10 MB.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image, style]
              properties:
                image:
                  type: string
                  format: binary
                roomType: { type: string, maxLength: 50 }
                style:
                  type: string
                  enum: [modern, minimal, classic, scandinavian, national, loft]
                note: { type: string, maxLength: 500 }
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DesignRequest'
        '400':
          description: Missing image, unsupported type, or image over 10 MB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '402':
          description: Daily request limit or daily AI budget exhausted (AC-11)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /design/requests/{id}:
    get:
      tags: [Design]
      security: [{ bearerAuth: [] }]
      summary: Status and variants for one request
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DesignRequest'
        '404':
          description: Unknown id, or the request belongs to another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /design/variants/{id}/materials:
    get:
      tags: [Design]
      summary: Material tags for a variant (orders module matches these to catalog)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DesignMaterial'